+++ src/main/java/nl/nn/testtool/run/ReportRunner.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2020, 2022 WeAreFrank!, 2018-2019 Nationale-Nederlanden
+   Copyright 2020, 2022-2023 WeAreFrank!, 2018-2019 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -21,6 +21,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import lombok.Setter;
+import nl.nn.testtool.Checkpoint;
 import nl.nn.testtool.MetadataExtractor;
 import nl.nn.testtool.Report;
 import nl.nn.testtool.SecurityContext;
@@ -32,25 +34,13 @@ import nl.nn.testtool.storage.StorageException;
  * @author Jaco de Groot
  */
 public class ReportRunner implements Runnable {
-	private TestTool testTool;
-	private SecurityContext securityContext;
+	private @Setter TestTool testTool;
+	private @Setter Storage debugStorage;
+	private @Setter SecurityContext securityContext;
 	private List<Report> reportsTodo = new ArrayList<Report>();
 	private int maximum = 1;
 	private Map<Integer, RunResult> results = Collections.synchronizedMap(new HashMap<Integer, RunResult>());
 	private boolean running = false;
-	private Storage debugStorage;
-
-	public void setTestTool(TestTool testTool) {
-		this.testTool = testTool;
-	}
-
-	public void setDebugStorage(Storage debugStorage) {
-		this.debugStorage = debugStorage;
-	}
-
-	public void setSecurityContext(SecurityContext securityContext) {
-		this.securityContext = securityContext;
-	}
 
 	public synchronized String run(List<Report> reports, boolean reset, boolean wait) {
 		if (running) {
@@ -109,8 +99,11 @@ public class ReportRunner implements Runnable {
 		return results;
 	}
 
-	public Report getRunResultReport(String runResultCorrelationId)
-			throws StorageException {
+	public Report getRunResultReport(String runResultCorrelationId) throws StorageException {
+		return getRunResultReport(debugStorage, runResultCorrelationId);
+	}
+
+	public static Report getRunResultReport(Storage storage, String runResultCorrelationId) throws StorageException {
 		Report report = null;
 		List<String> metadataNames = new ArrayList<String>();
 		metadataNames.add("storageId");
@@ -120,16 +113,53 @@ public class ReportRunner implements Runnable {
 		searchValues.add(runResultCorrelationId);
 		List<List<Object>> metadata = null;
 		// TODO in Reader.getMetadata kun je ook i < numberOfRecords veranderen in result.size() < numberOfRecords zodat je hier 1 i.p.v. -1 mee kunt geven maar als je dan zoekt op iets dat niet te vinden is gaat hij alle records door. misschien debugStorage.getMetadata een extra paremter geven, numberOfRecordsToConsider en numberOfRecordsToReturn i.p.v. numberOfRecords? (let op: logica ook in mem storage aanpassen)
-		metadata = debugStorage.getMetadata(-1, metadataNames, searchValues,
+		metadata = storage.getMetadata(-1, metadataNames, searchValues,
 				MetadataExtractor.VALUE_TYPE_OBJECT);
 		if (metadata != null && metadata.size() > 0) {
 			Integer runResultStorageId = (Integer)((List<Object>)metadata.get(0)).get(0);
-			report = debugStorage.getReport(runResultStorageId);
+			report = storage.getReport(runResultStorageId);
 		}
 		return report;
 	}
 
-	public Storage getDebugStorage() {
-		return debugStorage;
+	public static String getRunResultInfo(Report report, Report runResultReport) {
+		int stubbedOrig = 0;
+		for (Checkpoint checkpoint : report.getCheckpoints()) {
+			if (checkpoint.isStubbed()) {
+				stubbedOrig++;
+			}
+		}
+		int stubbedResult = 0;
+		int noStubInOriginalReport = 0;
+		int correlated = 0;
+		for (Checkpoint checkpoint : runResultReport.getCheckpoints()) {
+			if (checkpoint.isStubbed()) {
+				stubbedResult++;
+			}
+			if (checkpoint.getStubNotFound() != null) {
+				noStubInOriginalReport++;
+			}
+			if (checkpoint.isOriginalCheckpointFound()) {
+				correlated++;
+			}
+		}
+		int totalOrig = report.getCheckpoints().size();
+		int totalResult = runResultReport.getCheckpoints().size();
+		int total = totalOrig;
+		if (totalResult > totalOrig) {
+			total = totalResult;
+		}
+		String info = "(" + (report.getEndTime() - report.getStartTime()) + " >> "
+				+ (runResultReport.getEndTime() - runResultReport.getStartTime()) + " ms)"
+				+ " (" + stubbedOrig + "/" + totalOrig + " >> " + stubbedResult + "/" + totalResult + " stubbed)"
+				+ " (" + correlated + "/" + total + " correlated)";
+		if (noStubInOriginalReport > 0) {
+			info = info + " Stub message not found in original report for " + noStubInOriginalReport + " checkpoint";
+			if (noStubInOriginalReport > 1) {
+				info = info + "s";
+			}
+		}
+		return info;
 	}
+
 }